backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.2.0
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.120.0
h11==0.16.0
idna==3.11
motor==3.7.1
passlib==1.7.4
pip==25.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.15.3
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
